Add Owner Controller and move files to new folder
